Don't listen for onPlayBackEnded events when waiting to send UpNext data.
